Applied the Design Pattern Template to Inventory
I applied the DP Template to the Inventory class in order to be able to
create more and more inventories who will have different behaviors for
what type of inventories they are ( Chest, mob, etc ).

This implementation is in a DevOps purpose, in case we decide to
create other stuff in the game.